commit 9: move place of modifyViewDef method from Catalog.class to Alter.class
