+++ fe/src/main/java/org/apache/doris/persist/EditLog.java
@@ -239,7 +239,7 @@ public class EditLog {
                 }
                 case OperationType.OP_MODIFY_VIEW_DEF: {
                     AlterViewInfo info = (AlterViewInfo) journal.getData();
-                    catalog.replayModifyViewDef(info);
+                    catalog.getAlterInstance().replayModifyViewDef(info);
                 }
                 case OperationType.OP_RENAME_PARTITION: {
                     TableInfo info = (TableInfo) journal.getData();
